requirements.txt
langgraph
langchain 
langchain-google-genai
langsmith
faiss-cpu
sentence-transformers
pydantic
python-dotenv
notion-client
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client